+++ R/bravo.R
@@ -8,27 +8,27 @@
 #' @param quiet Logical. If TRUE, will display messages about potential missing census information. The default is FALSE.
 #' @param ... Arguments passed to \code{\link[tidycensus]{get_acs}} to select state, county, and other arguments for census characteristics
 #'
-#' @details This function will compute the spatial Educational Isolation Index (*EI*) of U.S. census tracts or counties for a specified geographical extent (e.g., the entire U.S. or a single state) based on Bravo et al. (2021) \doi{10.3390/ijerph18179384} who originally designed the metric for the educational isolation of individual without a college degree. This function provides the computation of *EI* for any of the U.S. Census Bureau educational attainment levels.
+#' @details This function will compute the spatial Educational Isolation Index (\emph{EI}) of U.S. census tracts or counties for a specified geographical extent (e.g., the entire U.S. or a single state) based on Bravo et al. (2021) \doi{10.3390/ijerph18179384} who originally designed the metric for the educational isolation of individual without a college degree. This function provides the computation of \emph{EI} for any of the U.S. Census Bureau educational attainment levels.
 #'
 #' The function uses the \code{\link[tidycensus]{get_acs}} to obtain U.S. Census Bureau 5-year American Community Survey characteristics used for the geospatial computation. The yearly estimates are available for 2009 onward when ACS-5 data are available but are available from other U.S. Census Bureau surveys. The five educational attainment levels (U.S. Census Bureau definitions) are:
 #' \itemize{
-#'  \item **B06009_002**: Less than high school graduate \code{'LtHS'}
-#'  \item **B06009_003**: High school graduate (includes equivalency) \code{'HSGiE'}
-#'  \item **B06009_004**: Some college or associate's degree \code{'SCoAD'}
-#'  \item **B06009_005**: Bachelor's degree \code{'BD'}
-#'  \item **B06009_006**: Graduate or professional degree \code{'GoPD'}
+#'  \item \strong{B06009_002}: Less than high school graduate \code{'LtHS'}
+#'  \item \strong{B06009_003}: High school graduate (includes equivalency) \code{'HSGiE'}
+#'  \item \strong{B06009_004}: Some college or associate's degree \code{'SCoAD'}
+#'  \item \strong{B06009_005}: Bachelor's degree \code{'BD'}
+#'  \item \strong{B06009_006}: Graduate or professional degree \code{'GoPD'}
 #' }
-#' Note: If \code{year = 2009}, then the ACS-5 data (2005-2009) are from the **B15002** question.
+#' Note: If \code{year = 2009}, then the ACS-5 data (2005-2009) are from the \strong{B15002} question.
 #'
-#' Use the internal \code{state} and \code{county} arguments within the \code{\link[tidycensus]{get_acs}} function to specify geographic extent of the data output. NOTE: Current version does not correct for edge effects (e.g., census geographies along the specified spatial extent border, coastline, or U.S.-Mexico / U.S.-Canada border) may have few neighboring census geographies, and *EI* values in these census geographies may be unstable. A stop-gap solution for the former source of edge effect is to compute the *EI* for neighboring census geographies (i.e., the states bordering a study area of interest) and then use the estimates of the study area of interest.
+#' Use the internal \code{state} and \code{county} arguments within the \code{\link[tidycensus]{get_acs}} function to specify geographic extent of the data output. NOTE: Current version does not correct for edge effects (e.g., census geographies along the specified spatial extent border, coastline, or U.S.-Mexico / U.S.-Canada border) may have few neighboring census geographies, and \emph{EI} values in these census geographies may be unstable. A stop-gap solution for the former source of edge effect is to compute the \emph{EI} for neighboring census geographies (i.e., the states bordering a study area of interest) and then use the estimates of the study area of interest.
 #'
-#' A census geography (and its neighbors) that has nearly all of its population with the specified educational attainment category (e.g., a Bachelor's degree or more) will have an *EI* value close to 1. In contrast, a census geography (and its neighbors) that is nearly none of its population with the specified educational attainment category (e.g., less than a Bachelor's degree) will have an *EI* value close to 0.
+#' A census geography (and its neighbors) that has nearly all of its population with the specified educational attainment category (e.g., a Bachelor's degree or more) will have an \emph{EI} value close to 1. In contrast, a census geography (and its neighbors) that is nearly none of its population with the specified educational attainment category (e.g., less than a Bachelor's degree) will have an \emph{EI} value close to 0.
 #'
 #' @return An object of class 'list'. This is a named list with the following components:
 #'
 #' \describe{
-#' \item{\code{ei}}{An object of class 'tbl' for the GEOID, name, *EI*, and raw census values of specified census geographies.}
-#' \item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute *EI*.}
+#' \item{\code{ei}}{An object of class 'tbl' for the GEOID, name, \emph{EI}, and raw census values of specified census geographies.}
+#' \item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute \emph{EI}.}
 #' }
 #'
 #' @import dplyr