+++ R/hoover.R
@@ -10,44 +10,44 @@
 #' @param quiet Logical. If TRUE, will display messages about potential missing census information. The default is FALSE.
 #' @param ... Arguments passed to \code{\link[tidycensus]{get_acs}} to select state, county, and other arguments for census characteristics
 #'
-#' @details This function will compute the aspatial Delta (*DEL*) of selected racial/ethnic subgroups and U.S. geographies for a specified geographical extent (e.g., the entire U.S. or a single state) based on Hoover (1941) \doi{10.1017/S0022050700052980} and Duncan, Cuzzort, and Duncan (1961; LC:60007089). This function provides the computation of *DEL* for any of the U.S. Census Bureau race/ethnicity subgroups (including Hispanic and non-Hispanic individuals).
+#' @details This function will compute the aspatial Delta (\emph{DEL}) of selected racial/ethnic subgroups and U.S. geographies for a specified geographical extent (e.g., the entire U.S. or a single state) based on Hoover (1941) \doi{10.1017/S0022050700052980} and Duncan, Cuzzort, and Duncan (1961; LC:60007089). This function provides the computation of \emph{DEL} for any of the U.S. Census Bureau race/ethnicity subgroups (including Hispanic and non-Hispanic individuals).
 #' 
-#' The function uses the \code{\link[tidycensus]{get_acs}} function to obtain U.S. Census Bureau 5-year American Community Survey characteristics used for the aspatial computation. The yearly estimates are available for 2009 onward when ACS-5 data are available (2010 onward for \code{geo_large = 'cbsa'}) but may be available from other U.S. Census Bureau surveys. The twenty racial/ethnic subgroups (U.S. Census Bureau definitions) are:
+#' The function uses the \code{\link[tidycensus]{get_acs}} function to obtain U.S. Census Bureau 5-year American Community Survey characteristics used for the aspatial computation. The yearly estimates are available for 2009 onward when ACS-5 data are available (2010 onward for \code{geo_large = 'cbsa'} and 2011 onward for \code{geo_large = 'csa'} or \code{geo_large = 'metro'}) but may be available from other U.S. Census Bureau surveys. The twenty racial/ethnic subgroups (U.S. Census Bureau definitions) are:
 #' \itemize{
-#'  \item **B03002_002**: not Hispanic or Latino \code{'NHoL'}
-#'  \item **B03002_003**: not Hispanic or Latino, white alone \code{'NHoLW'}
-#'  \item **B03002_004**: not Hispanic or Latino, Black or African American alone \code{'NHoLB'}
-#'  \item **B03002_005**: not Hispanic or Latino, American Indian and Alaska Native alone \code{'NHoLAIAN'}
-#'  \item **B03002_006**: not Hispanic or Latino, Asian alone \code{'NHoLA'}
-#'  \item **B03002_007**: not Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone \code{'NHoLNHOPI'}
-#'  \item **B03002_008**: not Hispanic or Latino, Some other race alone \code{'NHoLSOR'}
-#'  \item **B03002_009**: not Hispanic or Latino, Two or more races \code{'NHoLTOMR'}
-#'  \item **B03002_010**: not Hispanic or Latino, Two races including Some other race \code{'NHoLTRiSOR'}
-#'  \item **B03002_011**: not Hispanic or Latino, Two races excluding Some other race, and three or more races \code{'NHoLTReSOR'}
-#'  \item **B03002_012**: Hispanic or Latino \code{'HoL'}
-#'  \item **B03002_013**: Hispanic or Latino, white alone \code{'HoLW'}
-#'  \item **B03002_014**: Hispanic or Latino, Black or African American alone \code{'HoLB'}
-#'  \item **B03002_015**: Hispanic or Latino, American Indian and Alaska Native alone \code{'HoLAIAN'}
-#'  \item **B03002_016**: Hispanic or Latino, Asian alone \code{'HoLA'}
-#'  \item **B03002_017**: Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone \code{'HoLNHOPI'}
-#'  \item **B03002_018**: Hispanic or Latino, Some other race alone \code{'HoLSOR'}
-#'  \item **B03002_019**: Hispanic or Latino, Two or more races \code{'HoLTOMR'}
-#'  \item **B03002_020**: Hispanic or Latino, Two races including Some other race \code{'HoLTRiSOR'}
-#'  \item **B03002_021**: Hispanic or Latino, Two races excluding Some other race, and three or more races \code{'HoLTReSOR'}
+#'  \item \strong{B03002_002}: not Hispanic or Latino \code{'NHoL'}
+#'  \item \strong{B03002_003}: not Hispanic or Latino, white alone \code{'NHoLW'}
+#'  \item \strong{B03002_004}: not Hispanic or Latino, Black or African American alone \code{'NHoLB'}
+#'  \item \strong{B03002_005}: not Hispanic or Latino, American Indian and Alaska Native alone \code{'NHoLAIAN'}
+#'  \item \strong{B03002_006}: not Hispanic or Latino, Asian alone \code{'NHoLA'}
+#'  \item \strong{B03002_007}: not Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone \code{'NHoLNHOPI'}
+#'  \item \strong{B03002_008}: not Hispanic or Latino, Some other race alone \code{'NHoLSOR'}
+#'  \item \strong{B03002_009}: not Hispanic or Latino, Two or more races \code{'NHoLTOMR'}
+#'  \item \strong{B03002_010}: not Hispanic or Latino, Two races including Some other race \code{'NHoLTRiSOR'}
+#'  \item \strong{B03002_011}: not Hispanic or Latino, Two races excluding Some other race, and three or more races \code{'NHoLTReSOR'}
+#'  \item \strong{B03002_012}: Hispanic or Latino \code{'HoL'}
+#'  \item \strong{B03002_013}: Hispanic or Latino, white alone \code{'HoLW'}
+#'  \item \strong{B03002_014}: Hispanic or Latino, Black or African American alone \code{'HoLB'}
+#'  \item \strong{B03002_015}: Hispanic or Latino, American Indian and Alaska Native alone \code{'HoLAIAN'}
+#'  \item \strong{B03002_016}: Hispanic or Latino, Asian alone \code{'HoLA'}
+#'  \item \strong{B03002_017}: Hispanic or Latino, Native Hawaiian and Other Pacific Islander alone \code{'HoLNHOPI'}
+#'  \item \strong{B03002_018}: Hispanic or Latino, Some other race alone \code{'HoLSOR'}
+#'  \item \strong{B03002_019}: Hispanic or Latino, Two or more races \code{'HoLTOMR'}
+#'  \item \strong{B03002_020}: Hispanic or Latino, Two races including Some other race \code{'HoLTRiSOR'}
+#'  \item \strong{B03002_021}: Hispanic or Latino, Two races excluding Some other race, and three or more races \code{'HoLTReSOR'}
 #' }
 #' 
 #' Use the internal \code{state} and \code{county} arguments within the \code{\link[tidycensus]{get_acs}} function to specify geographic extent of the data output.
 #' 
-#' *DEL* is a measure of the proportion of members of one subgroup(s) residing in geographic units with above average density of members of the subgroup(s). The index provides the proportion of a subgroup population that would have to move across geographic units to achieve a uniform density. *DEL* can range in value from 0 to 1.
+#' \emph{DEL} is a measure of the proportion of members of one subgroup(s) residing in geographic units with above average density of members of the subgroup(s). The index provides the proportion of a subgroup population that would have to move across geographic units to achieve a uniform density. \emph{DEL} can range in value from 0 to 1.
 #' 
-#' Larger geographies available include state \code{geo_large = 'state'}, county \code{geo_large = 'county'}, Core Based Statistical Area \code{geo_large = 'cbsa'}, and census tract \code{geo_large = 'tract'} levels. Smaller geographies available include, county \code{geo_small = 'county'}, census tract \code{geo_small = 'tract'}, and census block group \code{geo_small = 'block group'} levels. If a larger geographical area is comprised of only one smaller geographical area (e.g., a U.S county contains only one census tract), then the *DEL* value returned is NA. If the larger geographical unit is Core Based Statistical Areas \code{geo_large = 'cbsa'}, only the smaller geographical units completely within a Core Based Statistical Area are considered in the *DEL* computation (see internal \code{\link[sf]{st_within}} function for more information) and recommend specifying all states within which the interested Core Based Statistical Areas are located using the internal \code{state} argument to ensure all appropriate smaller geographical units are included in the *DEL* computation.
+#' Larger geographies available include state \code{geo_large = 'state'}, county \code{geo_large = 'county'}, census tract \code{geo_large = 'tract'}, Core Based Statistical Area \code{geo_large = 'cbsa'}, Combined Statistical Area \code{geo_large = 'csa'}, and Metropolitan Division \code{geo_large = 'metro'} levels. Smaller geographies available include, county \code{geo_small = 'county'}, census tract \code{geo_small = 'tract'}, and census block group \code{geo_small = 'block group'} levels. If a larger geographical area is comprised of only one smaller geographical area (e.g., a U.S county contains only one census tract), then the \emph{DEL} value returned is NA. If the larger geographical unit is Combined Based Statistical Areas \code{geo_large = 'csa'} or Core Based Statistical Areas \code{geo_large = 'cbsa'}, only the smaller geographical units completely within a larger geographical unit are considered in the \emph{DEL} computation (see internal \code{\link[sf]{st_within}} function for more information) and recommend specifying all states within which the interested larger geographical unit are located using the internal \code{state} argument to ensure all appropriate smaller geographical units are included in the \emph{DEL} computation.
 #' 
 #' @return An object of class 'list'. This is a named list with the following components:
 #' 
 #' \describe{
-#' \item{\code{del}}{An object of class 'tbl' for the GEOID, name, and *DEL* at specified larger census geographies.}
+#' \item{\code{del}}{An object of class 'tbl' for the GEOID, name, and \emph{DEL} at specified larger census geographies.}
 #' \item{\code{del_data}}{An object of class 'tbl' for the raw census values at specified smaller census geographies.}
-#' \item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute *DEL*.}
+#' \item{\code{missing}}{An object of class 'tbl' of the count and proportion of missingness for each census variable used to compute \emph{DEL}.}
 #' }
 #' 
 #' @import dplyr
@@ -55,7 +55,7 @@
 #' @importFrom stats complete.cases
 #' @importFrom tidycensus get_acs
 #' @importFrom tidyr pivot_longer separate
-#' @importFrom tigris core_based_statistical_areas
+#' @importFrom tigris combined_statistical_areas core_based_statistical_areas metro_divisions
 #' @importFrom utils stack
 #' @export
 #' 
@@ -86,7 +86,7 @@ hoover <- function(geo_large = 'county',
                    ...) {
   
   # Check arguments
-  match.arg(geo_large, choices = c('state', 'county', 'tract', 'cbsa'))
+  match.arg(geo_large, choices = c('state', 'county', 'tract', 'cbsa', 'csa', 'metro'))
   match.arg(geo_small, choices = c('county', 'tract', 'block group'))
   stopifnot(is.numeric(year), year >= 2009) # all variables available 2009 onward
   match.arg(
@@ -210,12 +210,50 @@ hoover <- function(geo_large = 'county',
     del_data <- del_data %>%
       dplyr::mutate(
         oid = lapply(win_cbsa, function(x) { 
-          tmp <- dat_cbsa[x, 2] %>% sf::st_drop_geometry()
+          tmp <- dat_cbsa[x, 3] %>% sf::st_drop_geometry()
           lapply(tmp, function(x) { if (length(x) == 0) NA else x })
         }) %>% 
           unlist(),
         cbsa = lapply(win_cbsa, function(x) { 
           tmp <- dat_cbsa[x, 4] %>% sf::st_drop_geometry()
+          lapply(tmp, function(x) { if (length(x) == 0) NA else x })
+        }) %>% 
+          unlist()
+      ) %>% 
+      sf::st_drop_geometry()
+  }
+  if (geo_large == 'csa') {
+    stopifnot(is.numeric(year), year >= 2011) # CSAs only available 2011 onward
+    dat_csa <- suppressMessages(suppressWarnings(tigris::combined_statistical_areas(year = year)))
+    win_csa <- sf::st_within(del_data, dat_csa)
+    del_data <- del_data %>%
+      dplyr::mutate(
+        oid = lapply(win_csa, function(x) { 
+          tmp <- dat_csa[x, 2] %>% sf::st_drop_geometry()
+          lapply(tmp, function(x) { if (length(x) == 0) NA else x })
+        }) %>% 
+          unlist(),
+        csa = lapply(win_csa, function(x) { 
+          tmp <- dat_csa[x, 3] %>% sf::st_drop_geometry()
+          lapply(tmp, function(x) { if (length(x) == 0) NA else x })
+        }) %>% 
+          unlist()
+      ) %>% 
+      sf::st_drop_geometry()
+  }
+  if (geo_large == 'metro') {
+    stopifnot(is.numeric(year), year >= 2011) # Metro Divisions only available 2011 onward
+    dat_metro <- suppressMessages(suppressWarnings(tigris::metro_divisions(year = year)))
+    win_metro <- sf::st_within(del_data, dat_metro)
+    del_data <- del_data %>%
+      dplyr::mutate(
+        oid = lapply(win_metro, function(x) { 
+          tmp <- dat_metro[x, 4] %>% sf::st_drop_geometry()
+          lapply(tmp, function(x) { if (length(x) == 0) NA else x })
+        }) %>% 
+          unlist(),
+        metro = lapply(win_metro, function(x) { 
+          tmp <- dat_metro[x, 5] %>% sf::st_drop_geometry()
           lapply(tmp, function(x) { if (length(x) == 0) NA else x })
         }) %>% 
           unlist()
@@ -305,7 +343,30 @@ hoover <- function(geo_large = 'county',
       dplyr::mutate(GEOID = oid) %>%
       dplyr::select(GEOID, cbsa, DEL) %>%
       .[.$GEOID != 'NANA', ] %>%
-      dplyr::distinct(GEOID, .keep_all = TRUE)
+      dplyr::distinct(GEOID, .keep_all = TRUE) %>%
+      dplyr::filter(complete.cases(.))
+  }
+  if (geo_large == 'csa') {
+    del <- del_data %>%
+      dplyr::left_join(DELtmp, by = dplyr::join_by(oid)) %>%
+      dplyr::select(oid, csa, DEL) %>%
+      unique(.) %>%
+      dplyr::mutate(GEOID = oid) %>%
+      dplyr::select(GEOID, csa, DEL) %>%
+      .[.$GEOID != 'NANA', ] %>%
+      dplyr::distinct(GEOID, .keep_all = TRUE) %>%
+      dplyr::filter(complete.cases(.))
+  }
+  if (geo_large == 'metro') {
+    del <- del_data %>%
+      dplyr::left_join(DELtmp, by = dplyr::join_by(oid)) %>%
+      dplyr::select(oid, metro, DEL) %>%
+      unique(.) %>%
+      dplyr::mutate(GEOID = oid) %>%
+      dplyr::select(GEOID, metro, DEL) %>%
+      .[.$GEOID != 'NANA', ] %>%
+      dplyr::distinct(GEOID, .keep_all = TRUE) %>%
+      dplyr::filter(complete.cases(.))
   }
   
   del <- del %>%